.py and .CSV files
So the Python script is the script itself. The CSV is the file that the code cross-references VINs to once it's found a match. This CSV contains location data for where the car is in the junk yard. Once a match is found (tire size to car), it finds this car in this CSV and pastes all info from the row which contains the VIN number to a txt file.